+++ tests/test_password_auth_fails_with_wrong_password.c
@@ -13,7 +13,8 @@ int test(LIBSSH2_SESSION *session)
     int rc;
 
     const char *userauth_list =
-        libssh2_userauth_list(session, USERNAME, strlen(USERNAME));
+        libssh2_userauth_list(session, USERNAME,
+                              (unsigned int)strlen(USERNAME));
     if(userauth_list == NULL) {
         print_last_session_error("libssh2_userauth_list");
         return 1;
@@ -25,8 +26,10 @@ int test(LIBSSH2_SESSION *session)
         return 1;
     }
 
-    rc = libssh2_userauth_password_ex(session, USERNAME, strlen(USERNAME),
-                                      WRONG_PASSWORD, strlen(WRONG_PASSWORD),
+    rc = libssh2_userauth_password_ex(session, USERNAME,
+                                      (unsigned int)strlen(USERNAME),
+                                      WRONG_PASSWORD,
+                                      (unsigned int)strlen(WRONG_PASSWORD),
                                       NULL);
     if(rc == 0) {
         fprintf(stderr, "Password auth succeeded with wrong password\n");
