+++ tests/test_password_auth_fails_with_wrong_username.c
@@ -14,7 +14,8 @@ int test(LIBSSH2_SESSION *session)
     int rc;
 
     const char *userauth_list =
-        libssh2_userauth_list(session, WRONG_USERNAME, strlen(WRONG_USERNAME));
+        libssh2_userauth_list(session, WRONG_USERNAME,
+                              (unsigned int)strlen(WRONG_USERNAME));
     if(userauth_list == NULL) {
         print_last_session_error("libssh2_userauth_list");
         return 1;
@@ -27,8 +28,9 @@ int test(LIBSSH2_SESSION *session)
     }
 
     rc = libssh2_userauth_password_ex(session, WRONG_USERNAME,
-                                      strlen(WRONG_USERNAME), PASSWORD,
-                                      strlen(PASSWORD), NULL);
+                                      (unsigned int)strlen(WRONG_USERNAME),
+                                      PASSWORD,
+                                      (unsigned int)strlen(PASSWORD), NULL);
     if(rc == 0) {
         fprintf(stderr, "Password auth succeeded with wrong username\n");
         return 1;
